Friend status button now a componet
