--- src/components/ExampleComponent.tsx
@@ -1,7 +1,0 @@
-export default function ExampleComponent({ userId, myId }: { userId: string | undefined, myId: string | undefined }) {
-  return (
-    <div>
-
-    </div>
-  )
-}
