+++ src/pages/User.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import api from '../api.ts';
-import { Button } from '@nextui-org/react';
+import { Button, CircularProgress } from '@nextui-org/react';
 import { useAuth } from '../hooks/useAuth.ts';
-import ExampleComponent from '../components/ExampleComponent.tsx';
+import FriendStatusButton from '../components/FriendStatusButton.tsx';
 
 type receivedFriendRequests = {
   from_id: string
@@ -18,7 +18,7 @@ interface Group {
   [key: string]: unknown;
 }
 
-type User = {
+export type User = {
   firstName: string;
   lastName: string;
   userName: string;
@@ -38,85 +38,10 @@ export default function User() {
   const [loading, setLoading] = useState(true);
   const auth = useAuth();
   const currentId = auth?.user?.id;
-  const [requestState, setRequestState] = useState<string>("");
 
 
-  function sendFriendRequest() {
-    const data = {
-      from_id: currentId,
-      to_id: user?.id,
-    };
-
-    api
-      .post('friend-requests/send-friend-request', data)
-      .then(() => {
-        setRequestState("friend-request/sent")
-      })
-      .catch(err => console.error(err));
-  }
 
 
-  function makerequesState() {
-
-    // Estados: 
-    // Son amigos
-    // El me envio una request a mi
-    // la request esta pendiente
-    // no son amigos
-    
-    
-    if (currentId == undefined || id == undefined){
-      return
-    }else {
-      const data = {
-        from_id: currentId,
-        to_id: id,
-      };
-      console.log(data)
-      api
-        .post('friend-requests/get-request-status', data)
-        .then((res) => {
-          setRequestState(res.data.status)
-          
-        })
-        .catch(err => console.error(err));
-    }
-    
-
-
-  }
-
-  function buttonEngine() {
-    switch (requestState) {
-      case "friend-request/not-sent":
-        sendFriendRequest();
-        break;
-      case "friend-request/already-friends":
-        // Eliminar amigo Implementar en el futuro
-        break;
-      case "friend-request/sent":
-        // Pendiente, no hace nada, 
-        // Quizas en el futuro, eliminar solicitud
-        break;
-      case "friend-request/received":
-        break;
-    }
-    makerequesState();
-  }
-
-  function getButtonText(id: string): string {
-    console.log("requestState:", requestState)
-    switch (id) {
-      case "friend-request/sent":
-        return "Pendiente"
-      case "friend-request/received":
-        return "Aceptar solicitud"
-      case "friend-request/already-friends":
-        return "Quitar amigo"
-      default:
-        return "Agregar amigo"
-    }
-  }
 
   useEffect(() => {
     api.get(`users/profile/${id}`)
@@ -126,23 +51,17 @@ export default function User() {
       .catch((err) => {
         console.log(err);
       }
-  )}, [id]);
+    ).finally(()=> setLoading(false))
 
 
-  useEffect(() => {
-    
-      
-      makerequesState()
-      setLoading(false)
-    
+}, [id]);
 
-  }, [user]);
 
 
   return loading ? (
 
     <div>
-      Loading
+      <CircularProgress />
     </div>
   ) : (
     <div className="flex gap-5 flex-col">
@@ -160,8 +79,8 @@ export default function User() {
           ))}
         </h3>
       </div>
-      <Button onClick={() => buttonEngine()} color="primary">{getButtonText(requestState)}</Button>
-      <ExampleComponent userId={user?.id} myId={currentId} />
+      
+      <FriendStatusButton userId={user?.id} myId={currentId} />
     </div>
   );
 }
